Create sm.h
取地址应该用unsigned long类型，以避免在64位系统下出错。

+++ sm.h
@@ -26,7 +26,7 @@ extern "C" {
 #endif
 
 #ifndef offsetof
-#define offsetof(type, member) ((unsigned int) &((type *)0)->member)
+#define offsetof(type, member) ((unsigned long) &((type *)0)->member)
 #endif
 
 #ifndef container_of
